fix: corrige bugs de código e migra projeto para Maven com JavaFX
Correções de código (code review):
- Corrige seletor CSS #tiulo -> #titulo (título nunca recebia estilo)
- Remove CSS carregado duas vezes (DashboardView + Main)
- Adiciona null check no getResource para evitar NullPointerException
- Corrige typo no botão: 'Relatóris' -> 'Relatórios'
- Substitui double por BigDecimal no campo preco (precisão monetária)
- Extrai método criarCard() eliminando código triplicado nos cards
- DashboardView recebe valores do construtor em vez de hardcoded
- Remove objeto Bebida de teste e System.out.println do Main
- Remove classe Aplication.java (código morto, scaffold vazio)

Infraestrutura:
- Adiciona pom.xml com dependência javafx-controls 21.0.5 e plugin javafx-maven-plugin
- Atualiza .classpath para usar Maven classpath container (remove User Library manual)
- Atualiza .project com maven2Nature e maven2Builder
- Adiciona /target/ ao .gitignore

+++ src/br/com/distribuidora/model/Bebida.java
@@ -1,17 +1,19 @@
 package br.com.distribuidora.model;
 
+import java.math.BigDecimal;
+
 public class Bebida {
 
     private int id;
     private String nome;
     private String categoria;
-    private double preco;
+    private BigDecimal preco;
     private int estoque;
 
     public Bebida() {
     }
 
-    public Bebida(int id, String nome, String categoria, double preco, int estoque) {
+    public Bebida(int id, String nome, String categoria, BigDecimal preco, int estoque) {
         this.id = id;
         this.nome = nome;
         this.categoria = categoria;
@@ -43,11 +45,11 @@ public class Bebida {
         this.categoria = categoria;
     }
 
-    public double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
 
@@ -58,7 +60,7 @@ public class Bebida {
     public void setEstoque(int estoque) {
         this.estoque = estoque;
     }
-    
+
     public String descricao() {
         return nome + " - " + categoria + " - R$ " + preco + " - Estoque: " + estoque;
     }
